auth/requirements.txt
Flask==1.1.2
python-dotenv==0.14.0
Flask-Migrate==2.5.3
Flask-SQLAlchemy==2.4.4
PyMySQL==0.10.1
